+++ src/BootstrapAsset.php
@@ -17,9 +17,9 @@ use yii\web\AssetBundle;
  */
 class BootstrapAsset extends AssetBundle
 {
-    public $sourcePath = '@bower/bootstrap/dist';
+    public $sourcePath = '@vendor/twbs/bootstrap/dist/css';
 
     public $css = [
-        'css/bootstrap.css',
+        'bootstrap.css',
     ];
 }