+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 bootstrap5 extension Change Log
 2.0.51 under development
 ------------------------
 
+- Bug #83: Removed bower asset and replaced with twbs/bootstrap package (simialbi)
 - Bug #88: Fix ButtonDropdown JS error (stevekr, simialbi)
 - Bug #103: Fix `LinkPager::getPageRange` when `maxButtons` is 2 (mspirkov)